Be smarter with expanding strings for tr
When expanding for tr, we now just expand any ranges in the string
directly. This makes sure we don't accidentally remove duplicates that
are significant for String#tr.

We also make the handling for negation explicit in Ruby so the expand
logic only has to do the actual expanding.

Fixes #1871